Initial commit
The initial version finds contours, computes equivalent size,
and finds persistent bobs, classified as blocks.
There are three output files:
- blocks.csv: block index,simulation time,x,y,area
- fort.16: time,blob index,x,y  for each contour at each time step
- fort.17: time,blob index,x,y,area for each blob at each time step

There is a big problem with data that has continuous contours, which
are open: It does not know how to discard them.

It works great for data with simple blobs though, as demonstrated in
the example file.