Major overhaul:
This version should now correctly distinguish different contours.
The distinction between open and closed contours is still based
on the maximum length of the contour - too long -> open.
The centroid and equivalent radius (and thus area) are computed
with simple means; the data-weighted approach does not seem to
work as it should.

This approach is slow, as it goes over every contour point several times.
However, at least it works, and improvement can be saught from here.